requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.1.4
dropbox==11.36.2
python-dotenv==1.0.0
setuptools>=42.0.0
requests>=2.25.0